refactor: Utils 패키지 생성
