+++ src/main/java/cinebox/common/utils/SecurityUtil.java
@@ -1,4 +1,4 @@
-package cinebox.security;
+package cinebox.common.utils;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
