+++ src/main/java/cinebox/domain/payment/service/PaymentServiceImpl.java
@@ -14,6 +14,7 @@ import cinebox.common.exception.payment.InvalidPaymentStatusException;
 import cinebox.common.exception.payment.NotFoundPaymentException;
 import cinebox.common.exception.payment.NotPaidBookingException;
 import cinebox.common.exception.user.NoAuthorizedUserException;
+import cinebox.common.utils.SecurityUtil;
 import cinebox.domain.booking.entity.Booking;
 import cinebox.domain.booking.repository.BookingRepository;
 import cinebox.domain.payment.dto.PaymentRequest;
@@ -21,7 +22,6 @@ import cinebox.domain.payment.dto.PaymentResponse;
 import cinebox.domain.payment.entity.Payment;
 import cinebox.domain.payment.repository.PaymentRepository;
 import cinebox.domain.user.entity.User;
-import cinebox.security.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
